+++ src/main/java/com/cupid/jikting/member/controller/FileUploadController.java
@@ -1,11 +1,7 @@
 package com.cupid.jikting.member.controller;
 
-import com.amazonaws.services.s3.AmazonS3Client;
-import com.amazonaws.services.s3.model.ObjectMetadata;
-import com.cupid.jikting.common.error.ApplicationError;
-import com.cupid.jikting.common.error.FileUploadException;
+import com.cupid.jikting.member.service.FileUploadService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -20,26 +16,10 @@ import java.io.IOException;
 @RequestMapping("/files")
 public class FileUploadController {
 
-    private final AmazonS3Client amazonS3Client;
-
-    @Value("${cloud.aws.s3.bucket}")
-    private String bucket;
+    private final FileUploadService fileUploadService;
 
     @PostMapping("/upload")
     public ResponseEntity<String> uploadFile(@RequestParam("file") MultipartFile file) throws IOException {
-        validateExist(file);
-        String fileName = file.getOriginalFilename();
-        String fileUrl = "https://" + bucket + ".s3.ap-northeast-2.amazonaws.com/" + fileName;
-        ObjectMetadata metadata = new ObjectMetadata();
-        metadata.setContentType(file.getContentType());
-        metadata.setContentLength(file.getSize());
-        amazonS3Client.putObject(bucket, fileName, file.getInputStream(), metadata);
-        return ResponseEntity.ok().body(fileUrl);
-    }
-
-    private static void validateExist(MultipartFile file) {
-        if (file.isEmpty()) {
-            throw new FileUploadException(ApplicationError.FILE_NOT_EXIST);
-        }
+        return ResponseEntity.ok().body(fileUploadService.save(file));
     }
 }
